requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.19.0
mysql-connector-python==9.5.0
packaging==25.0
Werkzeug==3.1.3
